apps/plotly_examples/python-scripts/requirements.txt
numpy
pandas
pydantic
openai
plotly-express
openpyxl
python-pptx
requests
kaleido==0.1.0post1
statsmodels